Start builtins hierarchy
We can now nest builtins within a broader module.
For example under, "Special Functions" we can have "Orthogonal
Polynomials" and "Bessel Functions".
