+++ mathics/builtin/specialfns/__init__.py
@@ -1,0 +1,3 @@
+"""
+Special Functions
+"""
